chore: fix eslint & test
